Created eschelon reduction function and procedure

+++ src/Matrix/Matrix_scan.java
@@ -23,12 +23,12 @@ public class Matrix_scan {
                 break;
             }
             System.out.println("Masukkan matrix :");
-            double[][] data_isi = new double[row][collumn];
+            float[][] data_isi = new float[row][collumn];
             for(i=0;i<row;i++)
             {
                 for(j=0;j<collumn;j++)
                 {
-                    data_isi[i][j] = scan.nextDouble();
+                    data_isi[i][j] = scan.nextFloat();
                 }
             }
             Matrix.set_data(m, data_isi);
